Added idle_time to inductiva resources list
Also did a small code refactor. Moved some code from machine_group to the cli.
