+++ inductiva/resources/__init__.py
@@ -5,3 +5,4 @@ from . import machine_groups
 from .machine_groups import estimate_machine_cost
 from .machine_cluster import MPICluster
 from .machine_types import list_available_machines
+from .machine_groups import get
